refactor(aigateway): split the contract vocabulary module into a package
core/chat_parameters.py had grown to 643 lines against the 450-line limit the
plan binds to touched Python files. It crossed the limit at 463 and was accepted
with a documented deviation twice, so the overage was compounding rather than
settling.

Split along the seam the file already marked: the published vocabulary, the
construction-time errors and the frozen value objects go to _types, the six pure
derivations over them go to _algebra. The three _DISABLED_* reason strings stay
together in _types even though two are used only by the algebra -- they are one
published wire vocabulary sharing one explanatory comment block, and splitting
the group by usage would lose the comparison between them.

The module becomes a package whose __init__ re-exports the full surface, so
every existing import keeps working and no call site moves. The surface is
unchanged: the same 26 public names, verified against the previous revision, with
__all__ matching it exactly so no pydantic or auth-mode name leaks through.

Behaviour-preserving: no test was added, changed or removed, which is the signal
the split carried no behaviour with it.

Refs: OME-602
